+++ NEWS.md
@@ -1,3 +1,9 @@
+knitrdata 0.4.2
+---------------------------------------------------------------------
+
+- Changing to insertText in Rstudio addins to speed insertion
+- New non-interactive addin to create empty data chunk
+
 knitrdata 0.4.1
 ---------------------------------------------------------------------